Changes input field value to the input type syntax on input type change
