+++ custom-intro.html
@@ -115,6 +115,6 @@
       <label>Rendering Tree</label>
       <textarea id="math_tree" wrap="off" readonly></textarea>
     </form>
-    <script src="renderer-script.js" type="module"></script>
+    <script src="./javascript/custom-intro.js" type="module"></script>
   </p>
 </section>
