+++ demo.html
@@ -64,6 +64,6 @@ description: Demo.
       <label>Rendering Tree</label>
       <textarea id="math_tree" wrap="off" readonly></textarea>
     </form>
-    <script src="./../renderer-script.js" type="module"></script>
+    <script src="./../javascript/demo.js" type="module"></script>
   </body>
 </html>
