+++ javascript/custom-intro.js
@@ -1,0 +1,7 @@
+import { Initialize, inputTimeouts } from "./core.js";
+
+Initialize();
+
+(() => {
+  document.querySelector("#fromfmt").addEventListener("change", inputTimeouts);
+})();
